% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_submissions.R
\name{download_submissions}
\alias{download_submissions}
\title{Download the JSON file for the company submissions for the cik.}
\usage{
download_submissions(cik, useragent, cache = "./.cache")
}
\arguments{
\item{cik, }{Character. The CIK of the company.}

\item{useragent, }{Character. Required by SEC: "Name email@domain.com"}

\item{cache, }{Character. Temporary cache for storing the tickers as a
csv file.}
}
\value{
data.frame containing the downloaded submissions.
}
\description{
Download the JSON file for the company submissions for the cik.
}
