% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tickers.R
\name{download_tickers}
\alias{download_tickers}
\title{Download the JSON file of ticker to cik numbers from EDGAR}
\usage{
download_tickers(useragent, cache = "./.cache")
}
\arguments{
\item{useragent, }{Character. Required by SEC: "Name email@domain.com"}

\item{cache, }{Character. Temporary cache for storing the tickers as a
csv file.}
}
\description{
Download the JSON file of ticker to cik numbers from EDGAR
}
