% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filing.R
\name{filing_urls}
\alias{filing_urls}
\title{Build the URL for a filing's sgml file.}
\usage{
filing_urls(cik, accession)
}
\arguments{
\item{cik}{Character. The company CIK (with or without leading zeros).}

\item{accession}{Character. The accession number in dashed format
(e.g. "0000320193-23-000106").}
}
\value{
A character string URL.
}
\description{
Build the URL for a filing's sgml file.
}
