% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cik.R
\name{get_cik}
\alias{get_cik}
\title{Retrieve the CIK for the given ticker symbol.}
\usage{
get_cik(ticker, useragent, cache = "./.cache")
}
\arguments{
\item{ticker, }{Character. Ticker symbol.}

\item{useragent, }{Character. Required by SEC: "Name email@domain.com"}

\item{cache, }{Character. Temporary cache for storing the tickers as a
csv file.}
}
\value{
10 character cik string.
}
\description{
Retrieve the CIK for the given ticker symbol.
}
