% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filing.R
\name{get_filings}
\alias{get_filings}
\title{Retrieve an SGML filing from EDGAR}
\usage{
get_filings(cik, accession, useragent, cache = "./.cache")
}
\arguments{
\item{cik, }{Character. The company CIK.}

\item{accession, }{Character. The accession number in dashed format.}

\item{useragent, }{Character. Required by SEC: "Name email@domain.com".}

\item{cache, }{Character. Temporary cache for storing the tickers as a
csv file.}
}
\value{
A data frame with columns: sequence, filename, type, description, size.
}
\description{
Retrieve an SGML filing from EDGAR
}
