% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions.R
\name{get_submissions}
\alias{get_submissions}
\title{Retrieve the JSON document of company filing submissions to EDGAR}
\usage{
get_submissions(ticker, useragent, cache = "./.cache")
}
\arguments{
\item{ticker, }{Character. Ticker symbol of the company.}

\item{useragent, }{Character. Required by SEC: "Name email@domain.com"}

\item{cache, }{Character. Temporary cache for storing the tickers as a
csv file.}
}
\value{
data.frame, Character. The submission list in data frame format.
}
\description{
Retrieve the JSON document of company filing submissions to EDGAR
}
