% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submissions.R
\name{get_submissions}
\alias{get_submissions}
\title{Retrieve the company EDGAR submissions.}
\usage{
get_submissions(cik, useragent, cache = "./.cache")
}
\arguments{
\item{useragent, }{Character. Required by SEC: "Name email@domain.com"}

\item{cache, }{Character. Temporary cache for storing the tickers as a
csv file.}

\item{ticker, }{Character. Ticker symbol of the company.}
}
\value{
data.frame, Character. The submission list in data frame format.
}
\description{
Retrieve the company EDGAR submissions.
}
